+++ hgraph_unit_tests/_wiring/test_de_dupping_of_nodes.py
@@ -1,5 +1,5 @@
-from hgraph import graph, GraphBuilder, wire_graph
-from hgraph import const, debug_print
+from hgraph import graph, GraphBuilder, wire_graph, debug_print
+from hgraph import const
 
 
 def test_node_de_duping():
@@ -8,9 +8,10 @@ def test_node_de_duping():
         """The main graph."""
         c = const(1)
         d = const(1)
-        debug_print("c+d", c+d)
+        debug_print("c+d", c + d)
 
     from hgraph._wiring._wiring_node_instance import WiringNodeInstanceContext
+
     with WiringNodeInstanceContext():
         out = wire_graph(main)
     assert isinstance(out, GraphBuilder)
@@ -22,9 +23,10 @@ def test_node_de_duping_2():
     def main():
         """The main graph."""
         c = const(1)
-        debug_print("c+d", c+1)
+        debug_print("c+d", c + 1)
 
     from hgraph._wiring._wiring_node_instance import WiringNodeInstanceContext
+
     with WiringNodeInstanceContext():
         out = wire_graph(main)
     assert isinstance(out, GraphBuilder)