+++ src/hgraph/_operators/__init__.py
@@ -1,6 +1,7 @@
 from hgraph._operators._analytical_operators import *
 from hgraph._operators._apply import *
 from hgraph._operators._date_operators import *
+from hgraph._operators._debug_tools import *
 from hgraph._operators._flow_control import *
 from hgraph._operators._graph_operators import *
 from hgraph._operators._operators import *